+++ examples/src/bin/tmp.rs
@@ -3,7 +3,7 @@
 use algebraeon::{
     nzq::{integer::*, natural::*, rational::*},
     rings::{
-        number::natural::factored::factor,
+        number::natural::factorization::factor,
         polynomial::{multipoly::*, polynomial::Polynomial},
         structure::{elements::*, structure::*},
     },
@@ -15,7 +15,8 @@ fn main() {
         std::env::set_var("RUST_BACKTRACE", "1");
     }
 
-    let n = Natural::from_str("435879897345023450789532407859897340783457803453420987534278053978411111").unwrap();
+    let n = Natural::from_str("89754902345089754074587943578934757489025340982435043289079823459780598711")
+        .unwrap();
     println!("{}", n);
     println!("{:?}", factor(n));
 }
